refactoring & add logout api
